requirements.txt
jmcomic
pillow 